adk_training/module_13_code_analyst/requirements.txt
google-adk>=1.28.0
python-dotenv
llama-index-core
llama-index-embeddings-google-genai

# Web UI (dodatkowe)
fastapi
uvicorn
python-multipart
jinja2
